dune-sendmsg: bug corrected.

+++ programs/utils/dune-sendmsg.cpp
@@ -100,9 +100,9 @@ main(int argc, char** argv)
     IMC::MagneticField* tmsg = new IMC::MagneticField;
     msg = tmsg;
     tmsg->setSourceEntity(atoi(argv[4]));
-    tmsg->x = atoi(argv[5]);
-    tmsg->y = atoi(argv[6]);
-    tmsg->z = atoi(argv[7]);
+    tmsg->x = atof(argv[5]);
+    tmsg->y = atof(argv[6]);
+    tmsg->z = atof(argv[7]);
   }
   else if (strcmp(argv[3], "DataSanity") == 0)
   {
